Move SignDisclaimer component from Sign.tsx into its own file. Import the new component in Checkout/index and render it below ChekoutContent. Adjust styling in Sign.tsx.
